Convert to package and split modules
