+++ escape/__init__.py
@@ -1,0 +1,3 @@
+from .game import Game
+
+__all__ = ["Game"]
