+++ escape/data/daemon.log
