+++ escape/data/deep.node.log
