+++ escape/data/mem.fragment
