+++ escape/data/npc/daemon.dialog
