+++ escape/data/npc/dreamer.dialog
