+++ escape/data/npc/oracle.dialog
