+++ escape/data/npc/sysop.dialog
