+++ escape/data/npc/wanderer.dialog
