+++ escape/game.py
@@ -837,18 +837,3 @@ class Game:
                     break
             else:
                 self._output(f"Unknown command: {cmd}")
-
-
-def main(argv: list[str] | None = None):
-    import argparse
-
-    parser = argparse.ArgumentParser(description="Escape the Terminal")
-    parser.add_argument("--color", action="store_true", help="enable ANSI colors")
-    args = parser.parse_args(argv)
-
-    game_color = True if args.color else None
-    Game(use_color=game_color).run()
-
-
-if __name__ == '__main__':
-    main()
