+++ escape_the_terminal-0.1.0.dist-info/METADATA
@@ -1,0 +1,3 @@
+Metadata-Version: 2.1
+Name: escape-the-terminal
+Version: 0.1.0
