+++ escape_the_terminal-0.1.0.dist-info/entry_points.txt
@@ -1,0 +1,2 @@
+[console_scripts]
+escape-terminal = escape.cli:main
